CBTC : CBTCNavigatePhase with CBTC event
