+++ src/main/java/fr/sncf/osrd/train/TrainState.java
@@ -1,5 +1,10 @@
 package fr.sncf.osrd.train;
 
+import java.util.ArrayDeque;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import fr.sncf.osrd.infra.trackgraph.Detector;
 import fr.sncf.osrd.infra_state.SignalState;
@@ -7,13 +12,9 @@ import fr.sncf.osrd.simulation.Simulation;
 import fr.sncf.osrd.simulation.SimulationError;
 import fr.sncf.osrd.simulation.TimelineEvent;
 import fr.sncf.osrd.speedcontroller.SpeedDirective;
+import fr.sncf.osrd.train.phases.NavigatePhaseState;
 import fr.sncf.osrd.train.phases.PhaseState;
-import fr.sncf.osrd.train.phases.SignalNavigatePhase;
 import fr.sncf.osrd.utils.DeepComparable;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.ArrayDeque;
 
 public final class TrainState implements Cloneable, DeepComparable<TrainState> {
     static final Logger logger = LoggerFactory.getLogger(TrainState.class);
@@ -257,9 +258,10 @@ public final class TrainState implements Cloneable, DeepComparable<TrainState> {
 
     /** Add or update aspects constraint of a signal */
     public void setAspectConstraints(SignalState signalState) {
-        if (currentPhaseState.getClass() != SignalNavigatePhase.State.class)
-            throw new RuntimeException("Expected SignalNavigatePhase state");
-        var navigatePhase = (SignalNavigatePhase.State) currentPhaseState;
+        if (!(currentPhaseState instanceof NavigatePhaseState)) {
+            throw new RuntimeException("Expected NavigatePhaseState");
+        }
+        var navigatePhase = (NavigatePhaseState) currentPhaseState;
         navigatePhase.setAspectConstraints(signalState, this);
     }
